Don't crash if we see a <DIDL-Lite> element with unknown upnp:class
Right now, this happens if there is a Spotify playlist in the queue --
AVTransport returns elements like

  <DIDL-Lite ...>
    <item id="..." parentID="spotify%3aview%3amade-for-x" restricted="true">
      <dc:title>Discover Weekly</dc:title>
      <upnp:class>object.container.playlistContainer#playlistItem</upnp:class>
      <dc:creator>Spotify</dc:creator>
      <upnp:albumArtURI>...</upnp:albumArtURI>
      <desc id="cdudn" nameSpace="urn:schemas-rinconnetworks-com:metadata-1-0/">SA_RINCON3...</desc>
    </item>
  </DIDL-Lite>

where the <upnp:class> element is the source of the trouble.

SoCo works around this problem by dynamically creating class objects
in the DidlObject hierarchy. That freaks me out -- I'd much rather add
a handful of empty subclasses.